add: Project typing
